relocate checks; translate to STDOUT
